+++ src/components/Footer/Footer.module.css
@@ -5,7 +5,6 @@
   width: 100%;
   background: var(--color-gray-1000);
   color: white;
-  margin-top: 50px;
 }
 
 .footerContent {
