+++ src/components/Frontpage/PhotoWall.module.css
@@ -1,107 +1,6 @@
-.grid {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: var(--size-space-200);
-}
-
-/*
- * Each row's height comes from an aspect-ratio picked so the cells land on the
- * natural photo proportions (3:2 landscape, 2:3 portrait) instead of being
- * cropped into panoramas. Column fractions and the row ratio are paired so the
- * tiles line up at equal heights. Images cover their cell.
- */
-.row {
-  display: grid;
-  gap: var(--size-space-200);
-}
-
-/* Big landscape (3:2) beside a portrait (2:3): widths 9:4, row ratio 13:6. */
-.bigLeft {
-  grid-template-columns: 9fr 4fr;
-  aspect-ratio: 13 / 6;
-}
-
-.smallLeftBig {
-  grid-template-columns: 4fr 9fr;
-  aspect-ratio: 13 / 6;
-}
-
-/* Two landscape (3:2) halves side by side. */
-.twoEqual {
-  grid-template-columns: 1fr 1fr;
-  aspect-ratio: 3 / 1;
-}
-
-/* Two portrait (2:3) halves side by side. */
-.twoPortrait {
-  grid-template-columns: 1fr 1fr;
-  aspect-ratio: 4 / 3;
-}
-
-/* Big landscape beside two stacked landscapes: widths 2:1, row ratio 9:4. */
-.bigLeftStack {
-  grid-template-columns: 2fr 1fr;
-  aspect-ratio: 9 / 4;
-}
-
-.stackBigRight {
-  grid-template-columns: 1fr 2fr;
-  aspect-ratio: 9 / 4;
-}
-
-/* Single full-width landscape. */
-.full {
-  grid-template-columns: 1fr;
-  aspect-ratio: 3 / 2;
-}
-
-/* The column of two stacked images inside the stack patterns. */
-.stack {
-  display: grid;
-  grid-template-rows: 1fr 1fr;
-  gap: var(--size-space-200);
-  min-height: 0;
-}
-
-.tile {
-  position: relative;
-  overflow: hidden;
-  min-height: 0;
-}
-
-.photo {
-  width: 100%;
-  height: 100%;
-  display: block;
-  object-fit: cover;
-  transition: transform 0.3s ease;
-}
-
-.tile:hover .photo {
-  transform: scale(1.04);
-}
-
 /* Centers the spinner shown while the next page loads. */
 .spinner {
   display: flex;
   justify-content: center;
   padding: var(--size-space-400);
-}
-
-/* On small screens collapse everything into a single full-width column. */
-@media (max-width: 640px) {
-  .row {
-    grid-template-columns: 1fr;
-    aspect-ratio: auto;
-  }
-
-  .stack {
-    grid-template-rows: none;
-  }
-
-  .photo {
-    height: auto;
-    aspect-ratio: 3 / 2;
-  }
 }
