+++ src/components/Search/SearchResults.module.css
@@ -4,3 +4,10 @@
   color: var(--color-text-default-secondary);
   font-family: var(--typography-family-sans);
 }
+
+/* Centers the spinner shown while the next page loads. */
+.spinner {
+  display: flex;
+  justify-content: center;
+  padding: var(--size-space-400);
+}
