+++ src/routes/__root.module.css
@@ -1,5 +1,18 @@
+.page {
+  display: flex;
+  flex-direction: column;
+}
+
+.innerPage {
+  display: flex;
+  flex-direction: column;
+  min-height: 100dvh;
+}
+
 .main {
+  flex: 1;
   margin: 2rem;
+  padding-bottom: 50px;
 }
 
 .fullPage {
