+++ src/routes/motive.module.css
@@ -1,169 +1,52 @@
-.motiveImage {
-  width: 256px;
-  height: 172px;
-  box-shadow: 2px 4px 4px 2px rgba(0, 0, 0, 0.25);
-  background-color: white;
+.page {
   position: relative;
-}
-
-.motiveImage img {
-  margin: 10px auto;
-  position: absolute;
-  left: 3px;
-  bottom: -7px;
-}
-
-.motiveImage:hover {
-  cursor: pointer;
-}
-
-.flex {
   display: flex;
-  flex-direction: row;
+  flex-direction: column;
+  gap: var(--size-space-200);
+}
+
+.header {
+  display: flex;
+  flex-direction: column;
+  gap: var(--size-space-100);
+  margin-bottom: var(--size-space-800);
+}
+
+.title {
+  margin: 0;
+  font-family: var(--typography-family-sans);
+  font-size: var(--typography-heading-size-large);
+  font-weight: var(--typography-weight-bold);
+  color: var(--color-text-brand-secondary);
+}
+
+.subtitle {
+  display: flex;
   flex-wrap: wrap;
-  justify-content: center;
-  gap: 10px;
-  margin: 0px;
+  gap: var(--size-space-600);
+  font-family: var(--typography-family-sans);
+  font-size: var(--typography-body-size-medium);
+  font-weight: var(--typography-body-font-weight-regular);
+  color: var(--color-text-default-secondary);
 }
 
-.filling-empty-space-childs {
-  width: 300px;
-  padding: 0px;
-  margin: 20px;
+.titleSkeleton {
+  height: calc(var(--typography-heading-size-large) * 1.25);
+  width: 320px;
+  max-width: 70%;
+  border-radius: var(--size-radius-200);
 }
 
-.filterAndImages {
-  display: flex;
-  flex-direction: row;
-  flex: 1;
+.subtitleSkeleton {
+  height: calc(var(--typography-body-size-medium) * 1.25);
+  width: 240px;
+  max-width: 50%;
+  border-radius: var(--size-radius-200);
 }
 
-.backgroundFlex {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  width: 100%;
-  border-radius: 5px;
-}
-
-.imageHeader {
-  height: 150px;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
+.message {
   text-align: center;
-}
-
-.filterText {
-  text-align: center;
-}
-
-.imageFilter {
-  width: 500px;
-  height: 1000px;
-  display: grid;
-  grid-template-columns: 50fr 1fr;
-}
-
-.headerText {
-  font-size: 40px;
-  margin: 20px;
-  color: #7e7e7e;
-}
-
-.vl {
-  height: 95%;
-  width: 1px;
-  background-color: #000000;
-  justify-self: end;
-}
-
-.imageFlex {
-  position: relative;
-  max-width: 100%;
-}
-
-.bigFlex {
-  display: flex;
-  width: 100%;
-  justify-content: center;
-  align-items: center;
-}
-
-.leftArrow {
-  -webkit-filter: invert(1);
-  filter: invert(1);
-  margin: 0px;
-}
-
-.leftArrow:hover {
-  cursor: pointer;
-}
-
-.rightArrow {
-  -webkit-filter: invert(1);
-  filter: invert(1);
-  margin: 0px;
-}
-
-.rightArrow:hover {
-  cursor: pointer;
-}
-
-.imageContainer {
-  position: absolute;
-  bottom: 3px;
-  left: 3px;
-  background-color: #000000;
-  opacity: 0.3;
-  height: 43px;
-  width: 250px;
-}
-
-.imageTitle {
-  color: #ffffff;
-  font-size: 14px;
-}
-
-.hr {
-  background-color: "#000000";
-  height: 2px;
-  width: 95%;
-}
-
-.backButton {
-  position: absolute;
-  top: 35px;
-  left: 20px;
-  z-index: 10;
-
-  background: #ad2f33;
-  border: none;
-  border-radius: 50%;
-  width: 48px;
-  height: 48px;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  color: white;
-
-  transition: all 0.2s ease;
-  cursor: pointer;
-
-  outline: none;
-}
-
-.backButton:hover {
-  background: rgba(0, 0, 0, 0.9);
-  transform: scale(1.05);
-}
-
-.backButton:active {
-  transform: scale(0.95);
-}
-
-.backgroundFlex {
-  position: relative;
+  padding: var(--size-space-800);
+  color: var(--color-text-default-secondary);
+  font-family: var(--typography-family-sans);
 }
